[오수안] 슬라이더의 위치를 받ㄱ아올 ㅅ수 없다 왜일까? 왜? 왜?ㅇ ㅙ?

+++ HODOengine/HODOclient/SliderSoundScript.h
@@ -12,7 +12,7 @@ public:
 
 private:
 	bool _isSettingOn;
-	int _soundVolume;
+	float _soundVolume;
 
 	HDData::AudioSource* _audio;
 	HDData::SliderUI* _slider;
